--- a/app/Http/Controllers/Api/Serializer/ApiJsonSerializer.php
+++ b/app/Http/Controllers/Api/Serializer/ApiJsonSerializer.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Api\Serializer;
 
+use Illuminate\Support\Str;
 use League\Fractal\Pagination\CursorInterface;
 use League\Fractal\Pagination\PaginatorInterface;
 use League\Fractal\Resource\ResourceInterface;
@@ -81,8 +82,14 @@ class ApiJsonSerializer extends SerializerAbstract
 				foreach ($items as $item) {
 					if (!array_key_exists($resourceKey, $item)) continue;
 					$resourceId = $item[$resourceKey];
-					$this->relationships[$resourceKey][$resourceId][$includedResourceName][] = $item['id'];
-					$this->includes[$includedResourceName][$item['id']] = $item;
+					$pluralIncludedResourceName = Str::plural($includedResourceName);
+
+					if (!isset($this->relationships[$resourceKey][$resourceId][$pluralIncludedResourceName]) ||
+						!in_array($item['id'], $this->relationships[$resourceKey][$resourceId][$pluralIncludedResourceName], true)) {
+						// Don't duplicate relation id when parent resource was included multiple times
+						$this->relationships[$resourceKey][$resourceId][$pluralIncludedResourceName][] = $item['id'];
+					}
+					$this->includes[$pluralIncludedResourceName][$item['id']] = $item;
 				}
 			}
 		}