--- a/app/Http/Controllers/Api/Serializer/ApiJsonSerializer.php
+++ b/app/Http/Controllers/Api/Serializer/ApiJsonSerializer.php
@@ -78,19 +78,20 @@ class ApiJsonSerializer extends SerializerAbstract
 			if (empty ($includedResources)) continue;
 
 			foreach ($includedResources as $includedResourceName => $items) {
+				// TODO PLAT-971 don't convert to array, we want to have single item
 				if (array_key_exists('id', $items)) $items = [$items];
 				foreach ($items as $item) {
 					if (!array_key_exists($resourceKey, $item)) continue;
 					$resourceId = $item[$resourceKey];
-					// make name for list of included resources plural
-					$pluralIncludedResourceName = Str::plural($includedResourceName);
 
-					if (!isset($this->relationships[$resourceKey][$resourceId][$pluralIncludedResourceName]) ||
-						!in_array($item['id'], $this->relationships[$resourceKey][$resourceId][$pluralIncludedResourceName], true)) {
+					if (!isset($this->relationships[$resourceKey][$resourceId][$includedResourceName]) ||
+						!in_array($item['id'], $this->relationships[$resourceKey][$resourceId][$includedResourceName], true)) {
 						// Don't duplicate relation id when parent resource was included multiple times
-						$this->relationships[$resourceKey][$resourceId][$pluralIncludedResourceName][] = $item['id'];
+						$this->relationships[$resourceKey][$resourceId][$includedResourceName][] = $item['id'];
 					}
-					$this->includes[$pluralIncludedResourceName][$item['id']] = $item;
+
+					// make name for list of included resources plural
+					$this->includes[Str::plural($includedResourceName)][$item['id']] = $item;
 				}
 			}
 		}