--- a/resources/assets/js/admin/store/modules/taxonomyTerms.js
+++ b/resources/assets/js/admin/store/modules/taxonomyTerms.js
@@ -8,8 +8,8 @@ const namespaced = true;
 
 const state = {...initialState};
 
-const includeTag = (term, {tag}) => {
-	term.tag = tag[term.tag[0]];
+const includeTag = (term, {tags}) => {
+	term.tag = tags[term.tags[0]];
 	return term;
 };