--- a/resources/assets/js/admin/store/modules/taxonomyTerms.js
+++ b/resources/assets/js/admin/store/modules/taxonomyTerms.js
@@ -9,7 +9,7 @@ const namespaced = true;
 const state = {...initialState};
 
 const includeTag = (term, {tags}) => {
-	term.tag = tags[term.tags[0]];
+	term.tag = {...tags[term.tag[0]]};
 	return term;
 };