--- a/resources/assets/js/components/user/MyOrders.vue
+++ b/resources/assets/js/components/user/MyOrders.vue
@@ -88,7 +88,7 @@ export default {
 							return {
 								...order,
 								invoices: (order.invoices || []).map(invoiceId => invoices[invoiceId]),
-								studyBuddy: order.study_buddy ? included.study_buddy[order.study_buddy[0]] : {},
+								studyBuddy: order.study_buddy ? included.study_buddies[order.study_buddy[0]] : {},
 								payments: (order.payments || []).map(paymentId => payments[paymentId])
 							};
 						});