* **singular vs plural in include parameter is important!** - we validate if the relation exists on a model and we miss the relation when we use wrong include
* add `eagerLoadIncludes()` for `.filter` endpoint and endpoints using `nestedSets`
* we have to pass camelCase class names to Eloquent `->with()` method and also our includes validator was using them to check if a relation exists on a model

Results:
* http://platform.local/papi/v2/taxonomy_terms/byTaxonomy/5?include=tag 
```
|               | before | after |
|---------------|--------|-------|
| db queries    | 1305   | 4     |
| response time | 1402ms | 220ms |
```

* http://platform.local/papi/v2/annotations/.filter
```
|               | before | after |
|---------------|--------|-------|
| db queries    | 276    | 8     |
| response time | 334ms  | 74ms  |
```

Example: http://platform.local/papi/v2/taxonomy_terms/byTaxonomy/6?include=tag
```
{  
   "0":{  
      "id":7100,
      "description":"",
      "parent_id":null,
      "tag":[1707]
   },
   "included":{  
      "tags":{  
         "1707":{  
            "id":1707,
            "name":"zzz",
            "description":null,
            "color":null,
            "taxonomy_terms":7100,
            "0":[]
         }
      }
   }
}
```

To test:
* [ ] Content classifier
* [ ] Taxonomy editor
* [ ] Users list
* [ ] Tags list
* [ ] My orders - user facing